Add files via upload
This is a sample anti cheat script that takes screenshot every 30 seconds and sent it to a discord channel. The python script uses Selenium to take screenshot of multiple web pages and save them locally. It creates a list of URLSs, navigate to each URL. waits for the page to load completely, take a screenshot and time modules.